Enable FC SPI
Remove LIS3MDLOld. It's been unused since the GTXR board and was only used for reference.

Enabled support for SPI communication via FC. FC protocol isn't implemented in this commit, but the SPI initialization is done. Also put pull-up resistors on all NCS pins to ensure that we aren't trying to read/write to sensors before we enter main.